% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_slope.R
\name{test_beta_A_slope}
\alias{test_beta_A_slope}
\title{Wrapper for the treatment effect trajectory (beta_A) -----------------------}
\usage{
test_beta_A_slope(nknots, fit)
}
\description{
Wrapper for the treatment effect trajectory (beta_A) -----------------------
}
