% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_slope.R
\name{test_beta_mu_slope}
\alias{test_beta_mu_slope}
\title{Wrapper for the baseline trajectory (beta_mu) -------------------------------}
\usage{
test_beta_mu_slope(nknots, fit)
}
\description{
Wrapper for the baseline trajectory (beta_mu) -------------------------------
}
