% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTM-package.R
\docType{package}
\name{TTM-package}
\alias{TTM-package}
\title{TTM: Terminal Trend Models on the Retrospective Time Scale}
\description{
Fits retrospective-time longitudinal models for terminal trend analysis using weighted generalized estimating equations and related joint modeling tools. The package provides functions to estimate spline-based baseline and treatment effect trajectories on the retrospective time scale, obtain pointwise confidence intervals, and visualize fitted trajectories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TChang-W/TTM}
  \item Report bugs at \url{https://github.com/TChang-W/TTM/issues}
}

}
\author{
\strong{Maintainer}: Zhigang Li \email{zhigang.li@ufl.edu}

Authors:
\itemize{
  \item Tingchang Wang \email{twang5@ufl.edu}
}

}
\keyword{internal}
