% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_slope.R
\name{test_beta_slope}
\alias{test_beta_slope}
\title{IMPORTANT: This function is for TTM_linear object with B-spline only.
Compute the change in slope of the either baseline trajectory or treatment-effect
around a knot and test its significance.}
\usage{
test_beta_slope(new_times, fit, var_name)
}
\arguments{
\item{new_times}{Numeric vector of knot positions. Needs to be of length 3 where the
middle value is the knot and the other two values are slightly before and after the knot.}

\item{fit}{A fitted `TTM_linear()` object}

\item{var_name}{Either `"X0"` (baseline trajectory beta_mu) or
`"A"` (treatment effect trajectory beta_A)}
}
\value{
A data.frame with slope1, slope2, diff, se_diff, p_value
}
\description{
IMPORTANT: This function is for TTM_linear object with B-spline only.
Compute the change in slope of the either baseline trajectory or treatment-effect
around a knot and test its significance.
}
